// chatbox.js/src/api/graphql/model.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Time: any;
  Bytes: any;
  Int64: any;
  Any: any;
  Upload: any;
};
// ...
export type ChatboxMessage = {
  __typename?: 'ChatboxMessage';
  id: Scalars['ID'];
  createdAt: Scalars['Time'];
  author?: Maybe<ChatboxAgent>;
  bodyHtml: Scalars['String'];
};

export type ChatboxAgent = {
  __typename?: 'ChatboxAgent';
  name: Scalars['String'];
  avatarUrl: Scalars['String'];
};

export type Chatbox = {
  __typename?: 'Chatbox';
  preferences: ChatboxPreferences;
  messages: Array<ChatboxMessage>;
};

export type ChatboxPreferences = {
  __typename?: 'ChatboxPreferences';
  color: Scalars['String'];
  name: Scalars['String'];
  avatarUrl: Scalars['String'];
  twitterUrl: Scalars['String'];
  facebookUrl: Scalars['String'];
  publicEmail: Scalars['String'];
  instagramUrl: Scalars['String'];
  whatsappNumber: Scalars['String'];
  mastodonUrl: Scalars['String'];
  homepageUrl: Scalars['String'];
  branding: Scalars['Boolean'];
  welcomeMessage: Scalars['String'];
};
// ...
export type SendChatboxMessageInput = {
  projectId: Scalars['ID'];
  body: Scalars['String'];
};